[#1300] Move implementation of print functionality to platform abstraction
