+++ Cargo.toml
@@ -34,6 +34,7 @@ members = [
 
     "iceoryx2-pal/concurrency-sync",
     "iceoryx2-pal/posix/",
+    "iceoryx2-pal/print/",
     "iceoryx2-pal/configuration/",
     "iceoryx2-pal/testing/",
     "iceoryx2-pal/os-api",
@@ -99,6 +100,7 @@ iceoryx2-bb-trait-tests = { version = "0.8.999", path = "iceoryx2-bb/trait-tests
 iceoryx2-bb-testing = { version = "0.8.999", path = "iceoryx2-bb/testing/" }
 iceoryx2-pal-concurrency-sync = { version = "0.8.999", path = "iceoryx2-pal/concurrency-sync/", default-features = false }
 iceoryx2-pal-posix = { version = "0.8.999", path = "iceoryx2-pal/posix/" }
+iceoryx2-pal-print = { version = "0.8.999", path = "iceoryx2-pal/print/" }
 iceoryx2-pal-os-api = { version = "0.8.999", path = "iceoryx2-pal/os-api/" }
 iceoryx2-pal-configuration = { version = "0.8.999", path = "iceoryx2-pal/configuration/" }
 iceoryx2-pal-testing = { version = "0.8.999", path = "iceoryx2-pal/testing/" }
